requirements.txt
pillow~=11.1.0
Flask~=3.1.0
MarkupSafe~=3.0.2